+++ src/App.vue
@@ -1,16 +1,16 @@
 <template>
   <div id="app">
-    <HelloWorld msg="Welcome to Your Vue.js + TypeScript App"/>
+    <Index/>
   </div>
 </template>
 
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator'
-import HelloWorld from './components/HelloWorld.vue'
+import Index from './components/Index.vue'
 
 @Component({
   components: {
-    HelloWorld
+    Index
   }
 })
 export default class App extends Vue {}
